requirements.txt
discord.py
Pillow~=9.4.0
numpy~=1.24.2
git+https://github.com/Rapptz/asqlite
opencv-python~=4.7.0.72
luaparser~=3.2.1